Normalisation and cleaning code and includes

+++ srcs/tools/putline.c
@@ -6,25 +6,25 @@
 /*   By: toliver <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/12/06 17:22:18 by toliver           #+#    #+#             */
-/*   Updated: 2018/10/08 16:36:47 by bede-fre         ###   ########.fr       */
+/*   Updated: 2018/10/09 16:12:09 by bede-fre         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "doom.h"
 
-static void		errset1(int *err, float *x, int difinc1, int difinc2)
+static void	errset1(int *err, float *x, int difinc1, int difinc2)
 {
 	*err -= difinc1;
-	*x += difinc2;
+	*x += (float)difinc2;
 }
 
-static void		errset2(int *err, float *y, int difinc0, int difinc3)
+static void	errset2(int *err, float *y, int difinc0, int difinc3)
 {
 	*err += difinc0;
-	*y += difinc3;
+	*y += (float)difinc3;
 }
 
-void			ft_putline(t_vec a, t_vec b, t_img *img, int color)
+void		ft_putline(t_vec a, t_vec b, t_img *img, int color)
 {
 	int	difinc[4];
 	int	err;
@@ -39,7 +39,8 @@ void			ft_putline(t_vec a, t_vec b, t_img *img, int color)
 	while (42)
 	{
 		px_to_img(img, a.x, a.y, color);
-		if ((int)roundf(a.x) == (int)roundf(b.x) && (int)roundf(a.y) == (int)roundf(b.y))
+		if ((int)roundf(a.x) == (int)roundf(b.x) &&
+			(int)roundf(a.y) == (int)roundf(b.y))
 			break ;
 		err2 = err;
 		if (err2 > -difinc[0])
